fix: formatDateLabel uses short weekday in GroupSessionRSVPCard

+++ src/components/session/GroupSessionRSVPCard.tsx
@@ -17,7 +17,7 @@ function formatDateLabel(value: string): string {
   const date = new Date(`${value}T00:00:00`);
   if (Number.isNaN(date.getTime())) return value;
   return date.toLocaleDateString(undefined, {
-    weekday: 'long',
+    weekday: 'short',
     month: 'short',
     day: 'numeric',
   });
